% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_assets.R
\name{tm_assets}
\alias{tm_assets}
\title{Get all assets}
\usage{
tm_assets(token, ...)
}
\arguments{
\item{token}{A valid access token}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A data frame with asset information. Each row represents a single asset
and the columns represent specific asset properties.
}
\description{
Gets the complete list of available assets and returns it as a data frame.
}
\examples{
\dontrun{
token <- tm_token()

tm_assets(token)
}
}
