% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_structures.R
\name{tm_child_structures}
\alias{tm_child_structures}
\title{Get child structures by parent structure Id}
\usage{
tm_child_structures(token, parentId)
}
\description{
Gets the structures that have a parent defined by parentId and returns them
in a data frame.
}
