% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_structures.R
\name{tm_descendant_structures}
\alias{tm_descendant_structures}
\title{Get descendant structures by parent structure ID}
\usage{
tm_descendant_structures(token, parent_id)
}
\arguments{
\item{token}{A valid access token}

\item{parent_id}{Parent structure ID in UUID format}
}
\value{
A data frame with all descendant structures of \code{parent_id}.
}
\description{
Retrieves the entire structure subtree underneath \code{parent_id} and returns it as
a data frame.
}
\examples{
\dontrun{
token <- tm_token()
roots <- tm_root_structures(token)

# Get descendant subtree structure underneath second root structure
tm_descendant_structures(token, roots$structureId[2])

# Get descendant subtree by specific parent structure id
tm_descendant_structures(token, "4e58e3ca-e57d-47b5-8619-20d39626116e")
}
}
