% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_data.R
\name{tm_interpoloated_data}
\alias{tm_interpoloated_data}
\title{Tags}
\usage{
tm_interpoloated_data(
  token,
  tag_name,
  start_date,
  end_date,
  step = 1,
  type = "linear",
  shift = 0
)
}
\arguments{
\item{token}{A valid access token}

\item{tag_name}{Tag name}

\item{start_date}{POSIXct object}

\item{end_date}{POSIXct object}

\item{step}{Time between returned time series data points expressed in seconds}

\item{type}{Interpolation type which is either "linear" or "stepped"}

\item{shift}{Time series offset expressed in seconds}
}
\value{

}
\description{
Tags
}
\examples{
\dontrun{

}
}
