% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_assets.R
\name{tm_tags}
\alias{tm_tags}
\title{Get all tags}
\usage{
tm_tags(token, ...)
}
\arguments{
\item{token}{A valid access token}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A data frame with tag information. Each row represents a single tag
and the columns represent specific tag attributes.
}
\description{
Gets the complete list of available tags including their attributes
and returns it as a data frame.
}
\examples{
 \dontrun{
  token <- tm_token()

  tm_tags(token)
 }
}
