% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_data.R
\name{tm_ts_interpolated_data}
\alias{tm_ts_interpolated_data}
\title{Get interpolated time series data by tag name}
\usage{
tm_ts_interpolated_data(
  token,
  tag_name,
  start_date,
  end_date,
  step = 1,
  type = "linear",
  shift = 0,
  ...
)
}
\arguments{
\item{token}{A valid access token}

\item{tag_name}{Tag name}

\item{start_date}{POSIXct object. Start date of the time series}

\item{end_date}{POSIXct object. End date of the time series}

\item{step}{Time increment between returned observations expressed in seconds}

\item{type}{Interpolation type which is either "linear" or "stepped"}

\item{shift}{Time series offset expressed in seconds}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A list with two elements:
\itemize{
\item \code{tag}: A list with tag information and three elements:
\itemize{
\item \code{tagName}: Length one character vector.  \code{tag_name} of the original request
\item \code{shift}: Length one integer vector. \code{shift} of the original request
\item \code{interpolationType}: Length one character vector. \code{interpolation_type} of the original request
}
\item \code{timeSeries}: A data frame with time series data:
\itemize{
\item \code{index}:  POSIXct vector. Time index of each observation
\item \code{value}: Double vector. Value of each observation
}
}
}
\description{
Get interpolated time series data by tag name
}
\examples{
\dontrun{
token <- tm_token()
start <-  ymd_hms("2019-09-15T05:16:00Z")
end <- ymd_hms("2019-09-15T06:16:00Z")

tm_ts_interpolated_data(token, "BA:CONC.1", start, end, 60)
}
}
