% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendminer-package.R
\docType{package}
\name{trendminer-package}
\alias{trendminer}
\alias{trendminer-package}
\title{trendminer: R client for the TrendMiner API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Lemm \email{alex23lemm@gmail.com}

}
\keyword{internal}
